Commandline rom launcher
Commandline rom launcher tools to be used with any linux distro. in combination with Lutris.  Development is in very early stages and code may need to be changed to work for your specific needs.  Code contains a modified version of the lutris_bulk_adder script.  The original can be viewed from the below link.  Project also contains an empty pga.db file which is modified from Lutris's default db file in that it has a few columns appended to account for file extension, file location, etc.

https://github.com/hwangeug/lutris-bulk-adder